requirements.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.7
backcall==0.2.0
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
codecov==2.1.9
collective.checkdocs==0.2
colorama==0.4.3
coverage==5.3
crayons==0.4.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.6
django-chartit==0.2.9
django-chartjs==2.2.1
django-ckeditor==5.9.0
django-debug-toolbar==2.2
django-excel==0.0.10
django-heroku==0.3.1
django-import-export==2.3.0
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
docutils==0.16
et-xmlfile==1.0.1
flake8==3.8.3
fs==2.4.11
gease==0.0.8
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
importlib-metadata==2.0.0
ipdb==0.13.3
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.5.3
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jinja2-fsloader==0.3.0
kiwisolver==1.2.0
lml==0.0.9
lxml==4.5.2
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
moban==0.8.2
moban-jinja2-github==0.0.3
mypy-extensions==0.4.3
nose==1.3.7
numpy==1.19.2
odfpy==1.4.1
openpyxl==3.0.5
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.5
psycopg2==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pyecharts-jupyter-installer==0.0.3
pyexcel==0.6.4
pyexcel-ezodf==0.3.4
pyexcel-handsontable==0.0.2
pyexcel-io==0.5.20
pyexcel-ods3==0.5.3
pyexcel-pygal==0.0.2
pyexcel-webio==0.1.4
pyexcel-xls==0.5.9
pyexcel-xlsx==0.5.8
pyflakes==2.2.0
pygal==2.4.0
Pygments==2.6.1
pyparsing==2.4.7
python-dateutil==1.5
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
texttable==1.6.3
toml==0.10.1
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.2.0
